+++ langgraph/backend/app/main.py
@@ -4,7 +4,6 @@ from app.utils.debug import maybe_attach_pycharm
 from app.graphs.graph import build_main_app
 from app.nodes.supervisor import make_node_supervisor
 from app.nodes.problem_framing import make_node_problem_framing
-from app.nodes.hilp import make_node_hilp
 from app.agents.problem_framing.agent import build_agent as build_pf_agent
 from app.agents.translator.agent import build_agent as build_translation_agent
 from app.nodes.detect_language import make_node_detect_language
@@ -21,7 +20,6 @@ def build_app_for_cli():
         goto_after="supervisor",
     )
 
-    node_hilp = make_node_hilp()
     node_supervisor = make_node_supervisor()
 
     translation_agent = build_translation_agent()
@@ -33,7 +31,6 @@ def build_app_for_cli():
         detect_language_node=node_detect_language,
         supervisor_node=node_supervisor,
         problem_framing_node=node_problem_framing,
-        hilp_node=node_hilp,
         translation_node=translation_node,
     )